bump_version.sh: Commit before creating a tag
This ensures that the tag points to the commit that updated the cargo
files rather than trailing behind it by 1 commit.

Note that updating the TOML file mysteriously updates the lock file as
well, but with a delay. The "cargo update --workspace" call removes any
race conditions related to this.