bump_version.sh: Add instructions to undo script
Print how to undo the results of the script.
Note that this still leaves the modified cargo files in your
working directory. We could remove them by adding "--hard" to the reset
command, but that would discard any changes accidentally added to the
commit. Better to leave the changes and let the user decide what to do
with them.