Exit with success when user asked for help
Exit with code 0 when the user provided the flag `-h`.
Previously, this was treated as an error as a shortcut to
allow the program to exit early.

The implementation is probably overkill (just look at the diff!),
but it gets the job done without relying on janky solutions
like inspecting the error message.